requirements.txt
browserstack-local
paver
selenium
pytest
pytest-variables
pytest-selenium